refactor:  monolithic AuthForm with specific form components
