+++ src/app/(auth)/_components/index.ts
@@ -2,7 +2,6 @@
 export { AuthCard } from "./layout/auth-card";
 
 // Form components
-export { AuthForm } from "./forms/auth-form";
 export { FormInput } from "./forms/form-input";
 export { ResetPasswordForm } from "./forms/reset-password-form";
