+++ src/app/(auth)/log-in/page.tsx
@@ -3,11 +3,10 @@
 import { logIn } from "@/server/actions/auth";
 import { useRouter } from "next/navigation";
 import { toast } from "sonner";
-import { AuthForm } from "../_components/forms/auth-form";
-import { AuthCard } from "../_components/layout/auth-card";
+import { LoginForm } from "../_components/forms/login-form";
 import { useAuthAction } from "../_hooks/use-auth-action";
 
-export default function LoginForm() {
+export default function LoginPage() {
     const router = useRouter();
     const [, action] = useAuthAction(async (formData) => {
         const result = await logIn(formData);
@@ -25,17 +24,8 @@ export default function LoginForm() {
     });
 
     return (
-        <AuthCard
-            title="Welcome back"
-            description="Enter your email to sign in to your account"
-            alternateText="Don't have an account?"
-            linkText="Create an account"
-            linkHref="/sign-up"
-        >
-            <AuthForm
-                mode="login"
-                action={action}
-            />
-        </AuthCard>
+        <div className="flex min-h-screen items-center justify-center">
+            <LoginForm action={action} />
+        </div>
     );
 }
